Move all the helpers under SwaggerUiEngine scope
