+++ app/controllers/swagger_ui_engine/application_controller.rb
@@ -1,6 +1,6 @@
 module SwaggerUiEngine
   class ApplicationController < ActionController::Base
-    include AuthConfigParser
+    include SwaggerUiEngine::AuthConfigParser
 
     protect_from_forgery with: :exception
     layout false
