--- app/helpers/swagger_ui_defaults.rb
@@ -1,25 +1,0 @@
-module SwaggerUiDefaults
-  def default_swagger_url
-    'http://petstore.swagger.io/v2/swagger.json'
-  end
-
-  def default_doc_expansion
-    'none'
-  end
-
-  def default_json_editor
-    false
-  end
-
-  def default_model_rendering
-    'schema'
-  end
-
-  def default_request_headers
-    false
-  end
-
-  def default_validator_url
-    '//online.swagger.io/validator'
-  end
-end
