+++ app/helpers/swagger_ui_engine/auth_config_parser.rb
@@ -1,0 +1,19 @@
+module SwaggerUiEngine
+  module AuthConfigParser
+    def admin_username
+      configuration.admin_username
+    end
+
+    def admin_password
+      configuration.admin_password
+    end
+
+    def basic_authentication_enabled?
+      admin_username && admin_password
+    end
+
+    def configuration
+      SwaggerUiEngine.configuration
+    end
+  end
+end
